refactor(fragments): delete unnecessary methods
