+++ app/src/main/res/layout/fragment_game.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".GameFragment">
+    tools:context=".presentation.game.fragments.GameFragment">
 
     <GridView
         android:id="@+id/GameGrid"
